move inventoryLevelsEndDay gets to new file
