+++ index.js
@@ -31,8 +31,10 @@ app.put('/users/:id', db.updateUser);
 app.delete('/users/:id', db.deleteUser);
 */
 app.get('/inventory_items', db.getInventoryItems);
-app.get('/inventoryLevelsEndDay', db.getInventoryLevelsEndDayRecommended);
-app.get('/inventoryLevelsEndDayArrive', db.getInventoryLevelsEndDayRecordArrival);
+
+const dbInventoryLevels = require('./inventoryLevelsEndDay');
+app.get('/inventoryLevelsEndDay', dbInventoryLevels.getInventoryLevelsEndDayRecommended);
+app.get('/inventoryLevelsEndDayArrive', dbInventoryLevels.getInventoryLevelsEndDayRecordArrival);
 
 
 app.post('/whatSalesTogether', getWhatSalesTogether);
